created MenuHooks and DropdownSubmenu to DRY up code
